requirements.txt
aiohttp
heroku3
alphabet_detector
beautifulsoup4
bleach
nudepy
cachetools
lxml
wget
gtts
configparser
Python_ARQ
currencyconverter
emoji==1.7.0
jikanpy
pykeyboard
wikipedia
feedparser
future
git+https://github.com/Itz-fork/py-trans.git
git+https://github.com/python-telegram-bot/ptbcontrib.git
gitpython
markdown2
multicolorcaptcha
pillow
psutil
psycopg2-binary
PyYAML
pyrate-limiter
python-telegram-bot==13.13
faker
requests==2.26.0
spamwatch
speedtest-cli
SQLAlchemy==1.4.36
ujson
pycryptodome
base58
pytimeparse
Tgcrypto
pyrogram==1.4.16
gtts
telethon 
motor
odmantic
telegraph
pymongo
opencv-python-headless
googletrans==4.0.0rc1
gpytranslate==1.4.0
google-trans-new
hiredis	
pretty_errors	
requests_html	
ShazamAPI	
html-telegraph-poster	
qrcode
requests
nekos.py
fontTools
opencv-python-headless
odmantic
tzlocal
pendulum
fuzzysearch
motor
envparse
#profanity (Credits to Julia)
better_profanity
textblob 
secureme
cloudscraper
img2pdf
multicolorcaptcha
youtube_search_python==1.4.9
youtube_search
lyrics_extractor
future
emoji
lxml
wget
gtts
beautifulsoup4
requests
psycopg2-binary
feedparser
faker
hachoir
pykeyboard
pynewtonmath
aiofiles
ffmpeg-python
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp
Pillow
CurrencyConverter
jikanpy
speedtest-cli
coffeehouse
regex
bleach
wikipedia
markdown2
telethon
telegraph
heroku3
spamwatch
alphabet_detector
pyrate-limiter
cachetools
ujson
pretty_errors
pygithub
TgCrypto
PyDictionary
git+https://github.com/Sorrybabe/downloader
asyncio
cloudscraper
dateparser
pymongo
dnspython
secureme
apscheduler
search_engine_parser
emoji-country-flag
countryinfo
html2text
bs4
redis
pybase64
pySmartDL
bing_image_downloader
pytz
jikanpy
tswift
better_profanity
textblob 
fontTools
opencv-python-headless
odmantic
tzlocal
fuzzysearch
img2pdf
multicolorcaptcha
hiredis	
pretty_errors	
requests_html	
html-telegraph-poster	
qrcode
search_engine_parser
